# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/appkeys": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all application keys for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "List application keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResData-array_models_Appkey"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create an application key for the current user (omit `otp_code` if OTP is not enabled, omit `password` if OTP is enabled)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Create application key",
                "parameters": [
                    {
                        "description": "Name and permissions",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAppkeyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResData-models_CreateAppkeyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            }
        },
        "/user/appkeys/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an application key for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Delete application key",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Appkey ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            }
        },
...
        "/user/otp": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate a QR code for setting up OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Set up OTP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResData-models_SetupOtpResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Allow sign in without OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Disable OTP",
                "parameters": [
                    {
                        "description": "OTP code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Prevent logging in without OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Enable OTP",
                "parameters": [
                    {
                        "description": "OTP code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EnableOtpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResData-models_EnableOtpResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            }
        },
...
        "/user/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Revoke access to a session and invalidate its bearer tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Revoke session",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Appkey": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "last_used",
                "name",
                "permissions",
                "prefix"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-06-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
                },
                "last_used": {
                    "type": "string",
                    "example": "2021-06-01T00:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "My App"
                },
                "permissions": {
                    "$ref": "#/definitions/models.AppkeyPermissions"
                },
                "prefix": {
                    "type": "string",
                    "example": "1xauog0QhGi6MUzo"
                }
            }
        },
        "models.AppkeyPermissions": {
            "type": "object",
            "required": [
                "can_edit_user_info",
                "can_view_user_info"
            ],
            "properties": {
                "can_edit_user_info": {
                    "type": "boolean"
                },
                "can_view_user_info": {
                    "type": "boolean"
                }
            }
        },
...
        "models.CreateAppkeyInput": {
            "type": "object",
            "required": [
                "name",
                "otp_code",
                "password",
                "permissions"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My App"
                },
                "otp_code": {
                    "type": "string",
                    "example": "123456"
                },
                "password": {
                    "type": "string",
                    "example": "password1234"
                },
                "permissions": {
                    "$ref": "#/definitions/models.AppkeyPermissions"
                }
            }
        },
        "models.CreateAppkeyResponse": {
            "type": "object",
            "required": [
                "key"
            ],
            "properties": {
                "key": {
                    "type": "string",
                    "example": "Bearer ..."
                }
            }
        },
...
        "models.EnableOtpInput": {
            "type": "object",
            "required": [
                "otp_code"
            ],
            "properties": {
                "otp_code": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "models.EnableOtpResponse": {
            "type": "object",
            "required": [
                "recovery_codes"
            ],
            "properties": {
                "recovery_codes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "12345678",
                        "12345678",
                        "12345678",
                        "12345678",
                        "12345678"
                    ]
                }
            }
        },
        "models.ResData-array_models_Appkey": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Appkey"
                    }
                }
            }
        },
...
        "models.ResData-models_CreateAppkeyResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.CreateAppkeyResponse"
                }
            }
        },
        "models.ResData-models_EnableOtpResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.EnableOtpResponse"
                }
            }
        },
        "models.ResData-models_SetupOtpResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.SetupOtpResponse"
                }
            }
        },
...
        "models.ResErr": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message here..."
                }
            }
        },
        "models.ResMsg": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success message here..."
                }
            }
        },
...
        "models.SetupOtpResponse": {
            "type": "object",
            "required": [
                "qr_code",
                "secret"
            ],
            "properties": {
                "qr_code": {
                    "type": "string",
                    "example": "data:image/png;base64,..."
                },
                "secret": {
                    "type": "string",
                    "example": "JBSWY3DPEHPK3PXP"
                }
            }
        },
...
        "models.UserPasswordInput": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password1234"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter \"Bearer\" followed by a space and then your token, e.g. \"Bearer YOUR_TOKEN\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}